refactor: Organizes the Select files (#21589)
