+++ superset-frontend/src/components/Select/Select.stories.tsx
@@ -16,19 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import React, {
-  ReactNode,
-  useState,
-  useCallback,
-  useRef,
-  useMemo,
-} from 'react';
-import Button from 'src/components/Button';
+import React from 'react';
 import ControlHeader from 'src/explore/components/ControlHeader';
-import AsyncSelect, { AsyncSelectProps, AsyncSelectRef } from './AsyncSelect';
-import { SelectOptionsType, SelectOptionsTypePage } from './utils';
+import { SelectOptionsType, SelectProps } from './types';
 
-import Select, { SelectProps } from './Select';
+import Select from './Select';
 
 export default {
   title: 'Select',
@@ -326,239 +318,6 @@ PageScroll.story = {
     controls: {
       disable: true,
     },
-    knobs: {
-      disable: true,
-    },
-  },
-};
-
-const USERS = [
-  'John',
-  'Liam',
-  'Olivia',
-  'Emma',
-  'Noah',
-  'Ava',
-  'Oliver',
-  'Elijah',
-  'Charlotte',
-  'Diego',
-  'Evan',
-  'Michael',
-  'Giovanni',
-  'Luca',
-  'Paolo',
-  'Francesca',
-  'Chiara',
-  'Sara',
-  'Valentina',
-  'Jessica',
-  'Angelica',
-  'Mario',
-  'Marco',
-  'Andrea',
-  'Luigi',
-  'Quarto',
-  'Quinto',
-  'Sesto',
-  'Franco',
-  'Sandro',
-  'Alehandro',
-  'Johnny',
-  'Nikole',
-  'Igor',
-  'Sipatha',
-  'Thami',
-  'Munei',
-  'Guilherme',
-  'Umair',
-  'Ashfaq',
-  'Amna',
-  'Irfan',
-  'George',
-  'Naseer',
-  'Mohammad',
-  'Rick',
-  'Saliya',
-  'Claire',
-  'Benedetta',
-  'Ilenia',
-].sort();
-
-export const AsynchronousSelect = ({
-  fetchOnlyOnSearch,
-  withError,
-  withInitialValue,
-  responseTime,
-  ...rest
-}: AsyncSelectProps & {
-  withError: boolean;
-  withInitialValue: boolean;
-  responseTime: number;
-}) => {
-  const [requests, setRequests] = useState<ReactNode[]>([]);
-  const ref = useRef<AsyncSelectRef>(null);
-
-  const getResults = (username?: string) => {
-    let results: { label: string; value: string }[] = [];
-
-    if (!username) {
-      results = USERS.map(u => ({
-        label: u,
-        value: u,
-      }));
-    } else {
-      const foundUsers = USERS.filter(u => u.toLowerCase().includes(username));
-      if (foundUsers) {
-        results = foundUsers.map(u => ({ label: u, value: u }));
-      } else {
-        results = [];
-      }
-    }
-    return results;
-  };
-
-  const setRequestLog = (results: number, total: number, username?: string) => {
-    const request = (
-      <>
-        Emulating network request with search <b>{username || 'empty'}</b> ...{' '}
-        <b>
-          {results}/{total}
-        </b>{' '}
-        results
-      </>
-    );
-
-    setRequests(requests => [request, ...requests]);
-  };
-
-  const fetchUserListPage = useCallback(
-    (
-      search: string,
-      page: number,
-      pageSize: number,
-    ): Promise<SelectOptionsTypePage> => {
-      const username = search.trim().toLowerCase();
-      return new Promise(resolve => {
-        let results = getResults(username);
-        const totalCount = results.length;
-        const start = page * pageSize;
-        const deleteCount =
-          start + pageSize < totalCount ? pageSize : totalCount - start;
-        results = results.splice(start, deleteCount);
-        setRequestLog(start + results.length, totalCount, username);
-        setTimeout(() => {
-          resolve({ data: results, totalCount });
-        }, responseTime * 1000);
-      });
-    },
-    [responseTime],
-  );
-
-  const fetchUserListError = async (): Promise<SelectOptionsTypePage> =>
-    new Promise((_, reject) => {
-      reject(new Error('Error while fetching the names from the server'));
-    });
-
-  const initialValue = useMemo(
-    () => ({ label: 'Valentina', value: 'Valentina' }),
-    [],
-  );
-
-  return (
-    <>
-      <div
-        style={{
-          width: DEFAULT_WIDTH,
-        }}
-      >
-        <AsyncSelect
-          {...rest}
-          ref={ref}
-          fetchOnlyOnSearch={fetchOnlyOnSearch}
-          options={withError ? fetchUserListError : fetchUserListPage}
-          placeholder={fetchOnlyOnSearch ? 'Type anything' : 'AsyncSelect...'}
-          value={withInitialValue ? initialValue : undefined}
-        />
-      </div>
-      <div
-        style={{
-          position: 'absolute',
-          top: 32,
-          left: DEFAULT_WIDTH + 100,
-          height: 400,
-          width: 600,
-          overflowY: 'auto',
-          border: '1px solid #d9d9d9',
-          padding: 20,
-        }}
-      >
-        {requests.map((request, index) => (
-          <p key={`request-${index}`}>{request}</p>
-        ))}
-      </div>
-      <Button
-        style={{
-          position: 'absolute',
-          top: 452,
-          left: DEFAULT_WIDTH + 580,
-        }}
-        onClick={() => {
-          ref.current?.clearCache();
-          setRequests([]);
-        }}
-      >
-        Clear cache
-      </Button>
-    </>
-  );
-};
-
-AsynchronousSelect.args = {
-  allowClear: false,
-  allowNewOptions: false,
-  fetchOnlyOnSearch: false,
-  pageSize: 10,
-  withError: false,
-  withInitialValue: false,
-  tokenSeparators: ['\n', '\t', ';'],
-};
-
-AsynchronousSelect.argTypes = {
-  ...ARG_TYPES,
-  header: {
-    table: {
-      disable: true,
-    },
-  },
-  invertSelection: {
-    table: {
-      disable: true,
-    },
-  },
-  pageSize: {
-    defaultValue: 10,
-    control: {
-      type: 'range',
-      min: 10,
-      max: 50,
-      step: 10,
-    },
-  },
-  responseTime: {
-    defaultValue: 0.5,
-    name: 'responseTime (seconds)',
-    control: {
-      type: 'range',
-      min: 0.5,
-      max: 5,
-      step: 0.5,
-    },
-  },
-};
-
-AsynchronousSelect.story = {
-  parameters: {
     knobs: {
       disable: true,
     },
