+++ superset-frontend/src/explore/components/controls/SelectAsyncControl/index.tsx
@@ -20,8 +20,7 @@ import React, { useEffect, useState } from 'react';
 import { t, SupersetClient } from '@superset-ui/core';
 import ControlHeader from 'src/explore/components/ControlHeader';
 import { Select } from 'src/components';
-import { SelectProps } from 'src/components/Select/Select';
-import { SelectOptionsType } from 'src/components/Select/utils';
+import { SelectOptionsType, SelectProps } from 'src/components/Select/types';
 import { SelectValue, LabeledValue } from 'antd/lib/select';
 import withToasts from 'src/components/MessageToasts/withToasts';
 import { getClientErrorObject } from 'src/utils/getClientErrorObject';
